adicionando nova feature e alterando o css

+++ countdown.js
@@ -1,4 +1,4 @@
-async function getTimestamp() {    
+async function getTimestamp() {
     const res = await fetch('http://localhost:8080/countdown')
     const eventList = await res.json()
     return eventList
@@ -15,28 +15,34 @@ function handleEvents(eventList) {
             const hours = Math.floor((finalTime / 1000) / 60 / 60 % 24)
             const minutes = Math.floor((finalTime / 1000) / 60 % 60)
             const seconds = Math.floor((finalTime / 1000) % 60)
+            events += '<tr>'
             if (days > 0) {
-                events += `<p>${element.eventName} - ${days} Days : ${hours} Hours : ${minutes} Minutes : ${seconds} Seconds.</p>`
+                events += `<td class="event">${element.eventName} - ${days} Days : ${hours} Hours : ${minutes} Minutes : ${seconds} Seconds.</td>`
             } else if (hours > 0) {
-                events += `<p>${element.eventName} - ${hours} Hours : ${minutes} Minutes : ${seconds} Seconds.</p>`
+                events += `<td class="event"></td><td>${element.eventName} - ${hours} Hours : ${minutes} Minutes : ${seconds} Seconds.</td>`
             } else if (minutes > 0) {
-                events += `<p>${element.eventName} - ${minutes} Minutes : ${seconds} Seconds.</p>`
+                events += `<td class="event"><${element.eventName} - ${minutes} Minutes : ${seconds} Seconds.</td>`
             } else if (seconds > 0) {
-                events += `<p>${element.eventName} - ${seconds} Seconds.</p>`
+                events += `<td class="event">${element.eventName} - ${seconds} Seconds.</td>`
             } else {
-                events += `<p>${element.eventName} - has arrived.</p>`
+                events += `<td class="event">${element.eventName} - has arrived.</td>`
             }
+            events += '</tr>'
             if (days === 0 && hours === 0 && minutes === 0 && seconds === 0) {
-                alert(`${element.eventName} it's on!!!`)
+                Swal.fire({
+                    icon: 'success',
+                    title: 'Yay!',
+                    text: `${element.eventName} has arrived`
+                })
             }
-        });
+        })
     }
     document.getElementById("eventList").innerHTML = events;
 }
 
 async function countdown() {
     const eventList = await getTimestamp()
-    interval = setInterval(handleEvents, 500, eventList)
+    interval = setInterval(handleEvents, 1000, eventList)
 }
 
 const countdownForm = document.querySelector('.form')
@@ -49,10 +55,18 @@ document.getElementById('create').addEventListener('click', async function (even
     const formData = new FormData(countdownForm)
     const timestamp = new Date(formData.get('eventTime'))
     const eventName = formData.get('eventName')
-    if (timestamp < new Date()) {
-        alert('Wrong Date!')
-    } else if (eventName === "") {
-        alert('Invalid Name!')
+    if (eventName === "") {
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Invalid name!'
+        })
+    } else if (timestamp < new Date()) {
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Invalid Date!'
+        })
     } else {
         const obj = { eventName: eventName, eventTime: timestamp.getTime() }
         await fetch('http://localhost:8080/countdown', {
@@ -62,14 +76,75 @@ document.getElementById('create').addEventListener('click', async function (even
             },
             body: JSON.stringify(obj)
         })
+        const Toast = Swal.mixin({
+            toast: true,
+            position: 'top-end',
+            showConfirmButton: false,
+            timer: 3000,
+            timerProgressBar: true,
+        })
+        Toast.fire({
+            icon: 'success',
+            title: 'A new event has been created'
+        })
     }
+    document.getElementById('eventTime').value = defaultDate
     clearInterval(interval)
     countdown()
 })
 
-document.getElementById('delete').onclick = function () {
-    fetch('http://localhost:8080/countdown', {
-        method: 'DELETE'
+document.getElementById('share').addEventListener('click', async function (event) {
+    event.preventDefault()
+    const inputOptions = await getTimestamp()
+    const currentEvent = await Swal.fire({
+        title: 'Select an event',
+        input: 'select',
+        showCancelButton: true,
+        inputOptions: inputOptions.map(element => {
+            return element.eventName
+        })
     })
-}
+    const selectedElement = inputOptions[currentEvent.value]
+    const now = new Date(parseInt(selectedElement.eventTime))
+    if (currentEvent.isConfirmed) {
+        navigator.clipboard.writeText(`Hi! I would like to invite you to the ${selectedElement.eventName} that will be held on ${now}. I count on your presence, Bye!`)
+        const Toast = Swal.mixin({
+            toast: true,
+            position: 'top-end',
+            showConfirmButton: false,
+            timer: 3000,
+            timerProgressBar: true,
+        })
+        Toast.fire({
+            icon: 'success',
+            title: 'Message copied to the clipboard'
+        })
+    }
+})
+
+document.getElementById('delete').addEventListener('click', async function (event) {
+    event.preventDefault()
+    const result = await Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    })
+    if (result.isConfirmed) {
+        Swal.fire(
+            'Deleted!',
+            'All events have been cleared.',
+            'success'
+        )
+        await fetch('http://localhost:8080/countdown', {
+            method: 'DELETE'
+        })
+        clearInterval(interval)
+        countdown()
+    }
+})
+
 countdown()
